refactor : pharmacy 부분 로그인 처리 (재수정)
