+++ src/main/java/com/pharmquest/pharmquest/domain/pharmacy/web/controller/PharmacyController.java
@@ -2,13 +2,11 @@ package com.pharmquest.pharmquest.domain.pharmacy.web.controller;
 
 import com.pharmquest.pharmquest.domain.pharmacy.service.PharmacyCommandService;
 import com.pharmquest.pharmquest.domain.pharmacy.web.dto.PharmacyRequestDTO;
+import com.pharmquest.pharmquest.domain.token.JwtUtil;
 import com.pharmquest.pharmquest.domain.user.data.User;
 import com.pharmquest.pharmquest.global.apiPayload.ApiResponse;
 import com.pharmquest.pharmquest.global.apiPayload.code.status.SuccessStatus;
-import com.pharmquest.pharmquest.global.jwt.JwtUtil;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -28,14 +26,15 @@ public class PharmacyController {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "PHARMACY201", description = "약국을 마이페이지에 성공적으로 스크랩했습니다."),
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "PHARMACY202", description = "약국을 스크랩을 해제했습니다."),
     })
-    public ApiResponse<Void> scrapPharmacy(String authorizationHeader, @RequestBody @Valid PharmacyRequestDTO.ScrapDto request){
+    public ApiResponse<Void> scrapPharmacy(@RequestHeader(value = "Authorization") String authorizationHeader
+                                          ,@RequestBody @Valid PharmacyRequestDTO.ScrapDto request) {
 
         User user = jwtUtil.getUserFromHeader(authorizationHeader);
         Boolean scrap = pharmacyCommandService.scrapPharmacy(user, request.getPlaceId());
 
         if (scrap) { // scrap이 true면 스크랩 동작 성공, scrap이 false면 스크랩 취소 동작 성공
             return ApiResponse.noResultSuccess(SuccessStatus.PHARMACY_SCRAP);
-        }else{
+        } else {
             return ApiResponse.noResultSuccess(SuccessStatus.PHARMACY_UNSCRAP);
         }
     }
